+++ libs/enclave-vm/src/sidecar/__tests__/reference-sidecar.spec.ts
@@ -1,11 +1,4 @@
-import {
-  ReferenceSidecar,
-  SidecarLimitError,
-  ReferenceNotFoundError,
-  REFERENCE_CONFIGS,
-  REF_ID_PATTERN,
-  isReferenceId,
-} from '../index';
+import { ReferenceSidecar, SidecarLimitError, REFERENCE_CONFIGS, REF_ID_PATTERN, isReferenceId } from '../index';
 
 describe('ReferenceSidecar', () => {
   describe('store and retrieve', () => {